Fixed signed to unsigned comparison in mailer.cpp
This was generating compile errors. Should fix issue #83.

+++ src/mailer.cpp
@@ -2,12 +2,12 @@
  * @file       mailer.cpp
  * @brief      Defines types for sending emails
  * @author     Eddie Carle &lt;eddie@isatec.ca&gt;
- * @date       May 12, 2019
- * @copyright  Copyright &copy; 2019 Eddie Carle. This project is released under
+ * @date       July 29, 2020
+ * @copyright  Copyright &copy; 2020 Eddie Carle. This project is released under
  *             the GNU Lesser General Public License Version 3.
  */
 /*******************************************************************************
-* Copyright (C) 2019 Eddie Carle [eddie@isatec.ca]                             *
+* Copyright (C) 2020 Eddie Carle [eddie@isatec.ca]                             *
 *                                                                              *
 * This file is part of fastcgi++.                                              *
 *                                                                              *
@@ -93,7 +93,7 @@ void Fastcgipp::Mail::Mailer::handler()
                             m_line = "EHLO ";
                             m_line += m_origin + '\n';
                             if(m_socket.write(m_line.data(), m_line.size())
-                                    != static_cast<unsigned>(m_line.size()))
+                                    != static_cast<ssize_t>(m_line.size()))
                             {
                                 ERROR_LOG("Error sending EHLO command to SMTP "\
                                         "server.")
@@ -152,7 +152,7 @@ void Fastcgipp::Mail::Mailer::handler()
                                 m_line = "MAIL FROM: <";
                                 m_line += m_email.from + ">\n";
                                 if(m_socket.write(m_line.data(), m_line.size())
-                                        != static_cast<unsigned>(m_line.size()))
+                                        != static_cast<ssize_t>(m_line.size()))
                                 {
                                     ERROR_LOG("Error sending MAIL command to "\
                                             "SMTP server.")
@@ -183,7 +183,7 @@ void Fastcgipp::Mail::Mailer::handler()
                             m_line = "RCPT TO: <";
                             m_line += m_email.to + ">\n";
                             if(m_socket.write(m_line.data(), m_line.size())
-                                    != static_cast<unsigned>(m_line.size()))
+                                    != static_cast<ssize_t>(m_line.size()))
                             {
                                 ERROR_LOG("Error sending RCPT command to SMTP "\
                                         "server.")
@@ -237,7 +237,8 @@ void Fastcgipp::Mail::Mailer::handler()
                             for(const auto& chunk: m_email.body)
                                 if(m_socket.write(
                                             chunk.data.get(),
-                                            chunk.size) != chunk.size)
+                                            chunk.size) != static_cast<ssize_t>(
+                                                chunk.size))
                                 {
                                     ERROR_LOG("Error sending data chunk "\
                                             "to SMTP server.")
